requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Babel==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.25
email-validator==2.1.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
torch==2.0.0
transformers==4.35.2
sentence-transformers==2.2.2
openai==1.52.2
httpx==0.27.2
aiohttp==3.9.1
requests==2.31.0
